snakebin: re-write as a WSGI application
This is much cleaner, and also demonstrates a way to build WSGI web
applications with a simple WSGI framework like Falcon.
