+++ zerocloud/snakebin/part2/zapp.yaml
@@ -1,0 +1,32 @@
+project_type: python
+
+execution:
+  groups:
+    - name: "snakebin"
+      path: file://python2.7:python
+      args: "snakebin.py"
+      devices:
+      - name: python2.7
+      - name: stdout
+        content_type: message/http
+      - name: stdin
+      - name: input
+        path: swift://~/snakebin-store
+
+meta:
+  Version: ""
+  name: "snakebin"
+  Author-email: ""
+  Summary: ""
+
+help:
+  description: ""
+  args: []
+
+bundling: ["snakebin.py", "save_file.py", "get_file.py", "index.html"]
+
+dependencies: [
+    "falcon",
+    "six",
+    "mimeparse",
+]
