--- zerocloud/snakebin/zapp_part3.yaml
@@ -1,52 +1,0 @@
-# This describes the type of application. Bundling and deployment
-# behavior can vary between application types.
-project_type: python
-
-# This section describes the runtime behavior of your zapp: which
-# groups of nodes to create and which nexe to invoke for each.
-execution:
-
-  # Your application can consist of multiple groups. This is typically
-  # used for map-reduce style jobs. This is a list of groups, so
-  # remember to add "-" infront of each group name.
-  groups:
-
-      # Name of this group. This is used if you need to connect groups
-      # with each other.
-    - name: "snakebin"
-
-      # The NaCl executable (nexe) to run on the nodes in this group.
-      path: file://python2.7:python
-
-      # Command line arguments for the nexe.
-      args: "snakebin.py"
-
-      # Input and output devices for this group.
-      devices:
-      - name: python2.7
-      - name: stdin
-      - name: stdout
-        content_type: message/http
-      - name: input
-        path: swift://~/snakebin-store
-
-# Meta-information about your zapp.
-meta:
-  Version: ""
-  name: "snakebin"
-  Author-email: ""
-  Summary: ""
-
-help:
-  # Short description of your zapp. This is used for auto-generated
-  # help.
-  description: ""
-
-  # Help for the command line arguments. Each entry is a two-tuple
-  # with an option name and an option help text.
-  args: []
-
-# Files to include in your zapp. Your can use glob patterns here, they
-# will be resolved relative to the location of this file.
-bundling: ["snakebin.py", "save_file.py", "get_file.py", "index.html",
-           "search_mapper.py", "search_reducer.py"]
